feat(conflict): Rename and iterate the interface contract
