+++ submitqueue/extension/dependency/conflict/targetoverlap/BUILD.bazel
@@ -2,8 +2,8 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["tango.go"],
-    importpath = "github.com/uber/submitqueue/submitqueue/extension/dependency/conflict/tango",
+    srcs = ["targetoverlap.go"],
+    importpath = "github.com/uber/submitqueue/submitqueue/extension/dependency/conflict/targetoverlap",
     visibility = ["//visibility:public"],
     deps = [
         "//submitqueue/entity:go_default_library",
@@ -14,11 +14,12 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["tango_test.go"],
+    srcs = ["targetoverlap_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//submitqueue/entity:go_default_library",
         "//submitqueue/extension/conflict:go_default_library",
+        "//submitqueue/extension/dependency/resolver:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
